pallet_asset_conversion::pallet::dispatchables

Function remove_liquidity

pub fn remove_liquidity<T: Config>(
    asset1: Box<T::AssetKind>,
    asset2: Box<T::AssetKind>,
    lp_token_burn: T::Balance,
    amount1_min_receive: T::Balance,
    amount2_min_receive: T::Balance,
    withdraw_to: T::AccountId,
)

Allows you to remove liquidity by providing the lp_token_burn tokens that will be burned in the process. With the usage of amount1_min_receive/amount2_min_receive it’s possible to control the min amount of returned tokens you’re happy with.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::remove_liquidity.