pallet_asset_conversion::pallet::dispatchables

Function swap_exact_tokens_for_tokens

pub fn swap_exact_tokens_for_tokens<T: Config>(
    path: Vec<Box<T::AssetKind>>,
    amount_in: T::Balance,
    amount_out_min: T::Balance,
    send_to: T::AccountId,
    keep_alive: bool,
)

Swap the exact amount of asset1 into asset2. amount_out_min param allows you to specify the min amount of the asset2 you’re happy to receive.

AssetConversionApi::quote_price_exact_tokens_for_tokens runtime call can be called for a quote.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::swap_exact_tokens_for_tokens.