Type Alias pallet_asset_conversion::pallet::Module

pub type Module<T> = Pallet<T>;

👎Deprecated: use Pallet instead

Type alias to Pallet, to be used by construct_runtime.

Generated by pallet attribute macro.

Aliased Type

struct Module<T>(/* private fields */);

Implementations

impl<T: Config> Pallet<T>

Pallet’s callable functions.

pub fn create_pool( origin: OriginFor<T>, asset1: Box<T::AssetKind>, asset2: Box<T::AssetKind> ) -> DispatchResult

Creates an empty liquidity pool and an associated new lp_token asset (the id of which is returned in the Event::PoolCreated event).

Once a pool is created, someone may Pallet::add_liquidity to it.

pub fn add_liquidity( origin: OriginFor<T>, asset1: Box<T::AssetKind>, asset2: Box<T::AssetKind>, amount1_desired: T::Balance, amount2_desired: T::Balance, amount1_min: T::Balance, amount2_min: T::Balance, mint_to: T::AccountId ) -> DispatchResult

Provide liquidity into the pool of asset1 and asset2. NOTE: an optimal amount of asset1 and asset2 will be calculated and might be different than the provided amount1_desired/amount2_desired thus you should provide the min amount you’re happy to provide. Params amount1_min/amount2_min represent that. mint_to will be sent the liquidity tokens that represent this share of the pool.

Once liquidity is added, someone may successfully call Pallet::swap_exact_tokens_for_tokens successfully.

pub fn remove_liquidity( origin: OriginFor<T>, asset1: Box<T::AssetKind>, asset2: Box<T::AssetKind>, lp_token_burn: T::Balance, amount1_min_receive: T::Balance, amount2_min_receive: T::Balance, withdraw_to: T::AccountId ) -> DispatchResult

Allows you to remove liquidity by providing the lp_token_burn tokens that will be burned in the process. With the usage of amount1_min_receive/amount2_min_receive it’s possible to control the min amount of returned tokens you’re happy with.

pub fn swap_exact_tokens_for_tokens( origin: OriginFor<T>, path: Vec<Box<T::AssetKind>>, amount_in: T::Balance, amount_out_min: T::Balance, send_to: T::AccountId, keep_alive: bool ) -> DispatchResult

Swap the exact amount of asset1 into asset2. amount_out_min param allows you to specify the min amount of the asset2 you’re happy to receive.

AssetConversionApi::quote_price_exact_tokens_for_tokens runtime call can be called for a quote.

pub fn swap_tokens_for_exact_tokens( origin: OriginFor<T>, path: Vec<Box<T::AssetKind>>, amount_out: T::Balance, amount_in_max: T::Balance, send_to: T::AccountId, keep_alive: bool ) -> DispatchResult

Swap any amount of asset1 to get the exact amount of asset2. amount_in_max param allows to specify the max amount of the asset1 you’re happy to provide.

AssetConversionApi::quote_price_tokens_for_exact_tokens runtime call can be called for a quote.

impl<T: Config> Pallet<T>

pub fn get_reserves( asset1: T::AssetKind, asset2: T::AssetKind ) -> Result<(T::Balance, T::Balance), Error<T>>

Returns the balance of each asset in the pool. The tuple result is in the order requested (not necessarily the same as pool order).

pub fn quote_price_exact_tokens_for_tokens( asset1: T::AssetKind, asset2: T::AssetKind, amount: T::Balance, include_fee: bool ) -> Option<T::Balance>

Used by the RPC service to provide current prices.

pub fn quote_price_tokens_for_exact_tokens( asset1: T::AssetKind, asset2: T::AssetKind, amount: T::Balance, include_fee: bool ) -> Option<T::Balance>

Used by the RPC service to provide current prices.

pub fn quote( amount: &T::Balance, reserve1: &T::Balance, reserve2: &T::Balance ) -> Result<T::Balance, Error<T>>

Calculates the optimal amount from the reserves.

pub fn get_amount_out( amount_in: &T::Balance, reserve_in: &T::Balance, reserve_out: &T::Balance ) -> Result<T::Balance, Error<T>>

Calculates amount out.

Given an input amount of an asset and pair reserves, returns the maximum output amount of the other asset.

pub fn get_amount_in( amount_out: &T::Balance, reserve_in: &T::Balance, reserve_out: &T::Balance ) -> Result<T::Balance, Error<T>>

Calculates amount in.

Given an output amount of an asset and pair reserves, returns a required input amount of the other asset.