pallet_asset_conversion::pallet

Type Alias Pools

pub type Pools<T: Config> = StorageMap<_GeneratedPrefixForStoragePools<T>, Blake2_128Concat, T::PoolId, PoolInfo<T::PoolAssetId>, OptionQuery>;

Map from PoolAssetId to PoolInfo. This establishes whether a pool has been officially created rather than people sending tokens directly to a pool’s public account.

Storage type is StorageMap with key type T :: PoolId and value type PoolInfo < T :: PoolAssetId >.

Aliased Type

pub struct Pools<T: Config>(/* private fields */);