Trait pallet_asset_conversion::AssetConversionApi

pub trait AssetConversionApi<Block: BlockT, Balance, AssetBalance, AssetId>: Core<Block>where
    Balance: Codec + MaybeDisplay,
    AssetBalance: Balance,
    AssetId: Codec,{
    // Provided methods
    fn quote_price_tokens_for_exact_tokens(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        asset1: AssetId,
        asset2: AssetId,
        amount: AssetBalance,
        include_fee: bool
    ) -> Result<Option<Balance>, ApiError> { ... }
    fn quote_price_exact_tokens_for_tokens(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        asset1: AssetId,
        asset2: AssetId,
        amount: AssetBalance,
        include_fee: bool
    ) -> Result<Option<Balance>, ApiError> { ... }
    fn get_reserves(
        &self,
        __runtime_api_at_param__: <Block as BlockT>::Hash,
        asset1: AssetId,
        asset2: AssetId
    ) -> Result<Option<(Balance, Balance)>, ApiError> { ... }
}

This runtime api allows people to query the size of the liquidity pools and quote prices for swaps.

Provided Methods

fn quote_price_tokens_for_exact_tokens( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, asset1: AssetId, asset2: AssetId, amount: AssetBalance, include_fee: bool ) -> Result<Option<Balance>, ApiError>

Provides a quote for Pallet::swap_tokens_for_exact_tokens.

Note that the price may have changed by the time the transaction is executed. (Use amount_in_max to control slippage.)

fn quote_price_exact_tokens_for_tokens( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, asset1: AssetId, asset2: AssetId, amount: AssetBalance, include_fee: bool ) -> Result<Option<Balance>, ApiError>

Provides a quote for Pallet::swap_exact_tokens_for_tokens.

Note that the price may have changed by the time the transaction is executed. (Use amount_out_min to control slippage.)

fn get_reserves( &self, __runtime_api_at_param__: <Block as BlockT>::Hash, asset1: AssetId, asset2: AssetId ) -> Result<Option<(Balance, Balance)>, ApiError>

Returns the size of the liquidity pool for the given asset pair.